# backend/openapi/schema.yaml
openapi: 3.0.3
info:
  title: Chemistry Apps API
  version: 1.0.0
  description: API de plataforma científica modular para ejecutar jobs asíncronos
    por plugins (calculator, random-numbers y molar-fractions), con observabilidad
    de progreso y logs en tiempo real, cache por hash y recuperación activa automática.
  contact:
    name: Chemistry Apps Team
    url: https://github.com/chem-gl/chemistry-apps
  license:
    name: MIT
    url: https://opensource.org/license/mit/
paths:
...
  /api/jobs/:
    get:
      operationId: jobs_list
      description: Devuelve el listado de jobs científicos registrados en todas las
        apps, incluyendo su estado actual. Permite filtros opcionales por plugin y
        estado.
      summary: Listar Jobs Científicos
      parameters:
      - in: query
        name: plugin_name
        schema:
          type: string
        description: Filtra por nombre de plugin/app científica.
      - in: query
        name: status
        schema:
          type: string
        description: 'Filtra por estado del job. Valores válidos: pending, running,
          paused, completed, failed.'
      tags:
      - Jobs
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScientificJob'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                EstadoInválido:
                  value:
                    detail: 'Invalid status filter. Allowed values are: pending, running,
                      paused, completed, failed.'
                  summary: Estado inválido
          description: Filtro de estado inválido.
    post:
      operationId: jobs_create
      description: Evalua reglas de hashing/caching y despacha al worker Celery solo
        cuando el job no tiene cache disponible.
      summary: Despachar Job Científico
      tags:
      - Jobs
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobCreateRequest'
            examples:
              EjemploDeCreaciónDeJob:
                value:
                  plugin_name: calculator
                  version: '1.0'
                  parameters:
                    op: add
                    a: 5
                    b: 3
                summary: Ejemplo de Creación de Job
                description: Creación de una tarea matemática en background.
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/JobCreateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/JobCreateRequest'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScientificJob'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                ErrorPluginNoRegistrado:
                  value:
                    detail: Plugin requested is not registered
                  summary: Error plugin no registrado
          description: Error de validacion en parametros.
  /api/jobs/{id}/:
    get:
      operationId: jobs_retrieve
      description: Recupera todo el estado guardado incluyendo outputs una vez finalizado
        el trabajo subyacente.
      summary: Consultar Estado y Resultados de un Job
      parameters:
      - in: path
        name: id
        schema:
          type: string
        description: UUID del job cientifico a consultar.
        required: true
      tags:
      - Jobs
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScientificJob'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: El trabajo no pudo ser encontrado en el Registry local.
  /api/jobs/{id}/cancel/:
    post:
      operationId: jobs_cancel_create
      description: Cancela un job en estado pending, running o paused de forma inmediata
        e irreversible. No es posible reactivar un job cancelado. Los jobs en estado
        completed, failed o cancelled no pueden ser cancelados.
      summary: Cancelar un Job (irreversible)
      parameters:
      - in: path
        name: id
        schema:
          type: string
        description: UUID del job científico.
        required: true
      tags:
      - Jobs
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobControlActionResponse'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                JobYaFinalizado:
                  value:
                    detail: 'No es posible cancelar un job en estado terminal (estado
                      actual: completed).'
                  summary: Job ya finalizado
          description: El job está en estado terminal y no puede cancelarse.
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Job no encontrado.
  /api/jobs/{id}/events/:
    get:
      operationId: jobs_events_retrieve
      description: Abre un stream Server-Sent Events para recibir actualizaciones
        de progreso en tiempo real hasta completar/fallar o alcanzar timeout.
      summary: Suscribirse a eventos de progreso de un Job (SSE)
      parameters:
      - in: path
        name: id
        schema:
          type: string
        description: UUID del job científico.
        required: true
      - in: query
        name: timeout_seconds
        schema:
          type: integer
        description: 'Tiempo máximo del stream SSE en segundos. Valor por defecto:
          30.'
      tags:
      - Jobs
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            text/event-stream:
              schema:
                type: string
              examples:
                EventoSSE:
                  value: |+
                    id: 4
                    event: job.progress
                    data: {"job_id":"8ca8c1fa-1f2f-4a13-9038-9e1be7d0ce24","status":"running","progress_percentage":35,"progress_stage":"running","progress_message":"Ejecutando plugin científico.","progress_event_index":4,"updated_at":"2026-03-10T12:00:00Z"}

                  summary: Evento SSE
          description: Flujo SSE de eventos job.progress con id incremental.
        '404':
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Job no encontrado.
  /api/jobs/{id}/logs/:
    get:
      operationId: jobs_logs_retrieve
      description: Devuelve eventos de log persistidos para un job en orden ascendente
        por event_index, con soporte de cursor incremental y límite de página.
      summary: Consultar historial de logs de un Job
      parameters:
      - in: query
        name: after_event_index
        schema:
          type: integer
        description: Retorna eventos con índice mayor al valor indicado.
      - in: path
        name: id
        schema:
          type: string
        description: UUID del job científico.
        required: true
      - in: query
        name: limit
        schema:
          type: integer
        description: Cantidad máxima de eventos a devolver. Máximo 500.
      tags:
      - Jobs
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobLogList'
              examples:
                ListadoDeLogsPorJob:
                  value:
                    job_id: 8ca8c1fa-1f2f-4a13-9038-9e1be7d0ce24
                    count: 2
                    next_after_event_index: 13
                    results:
                    - job_id: 8ca8c1fa-1f2f-4a13-9038-9e1be7d0ce24
                      event_index: 12
                      level: info
                      source: random_numbers.plugin
                      message: Procesando lote de generación.
                      payload:
                        current_batch_size: 5
                      created_at: '2026-03-11T10:25:00Z'
                    - job_id: 8ca8c1fa-1f2f-4a13-9038-9e1be7d0ce24
                      event_index: 13
                      level: debug
                      source: random_numbers.plugin
                      message: Número generado correctamente.
                      payload:
                        generated_number: 1024
                      created_at: '2026-03-11T10:25:01Z'
                  summary: Listado de logs por job
                  description: Página de logs de ejecución de un job.
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Job no encontrado.
  /api/jobs/{id}/logs/events/:
    get:
      operationId: jobs_logs_events_retrieve
      description: Abre un stream Server-Sent Events para recibir logs en tiempo real
        hasta timeout o finalización del job.
      summary: Suscribirse a logs de un Job (SSE)
      parameters:
      - in: path
        name: id
        schema:
          type: string
        description: UUID del job científico.
        required: true
      - in: query
        name: timeout_seconds
        schema:
          type: integer
        description: 'Tiempo máximo del stream SSE en segundos. Valor por defecto:
          30.'
      tags:
      - Jobs
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            text/event-stream:
              schema:
                type: string
              examples:
                EventoSSEDeLog:
                  value: |+
                    id: 12
                    event: job.log
                    data: {"job_id":"8ca8c1fa-1f2f-4a13-9038-9e1be7d0ce24","event_index":12,"level":"info","source":"random_numbers.plugin","message":"Procesando lote de generación.","payload":{"current_batch_size":5},"created_at":"2026-03-11T10:25:00Z"}

                  summary: Evento SSE de log
          description: Flujo SSE de eventos job.log con id incremental.
        '404':
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Job no encontrado.
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      description: Schema consistente para respuestas de error API.
      properties:
        detail:
          type: string
          description: Descripción del error manejado por API.
      required:
      - detail
    JobControlActionResponse:
      type: object
      description: Respuesta estándar para acciones de control (pause/resume).
      properties:
        detail:
          type: string
          description: Resultado de la operación de control.
        job:
          description: Snapshot actualizado del job tras la acción.
      required:
      - detail
      - job
    JobCreateRequest:
      type: object
      description: |-
        Serializer puro para las peticiones de creación de un nuevo
        Job científico, siguiendo fuertemente el schema de OpenAPI.
      properties:
        plugin_name:
          type: string
          minLength: 1
          description: El nombre registrado del plugin o algoritmo a ejecutar.
          maxLength: 100
        version:
          type: string
          minLength: 1
          default: '1.0'
          description: Versión asociada del algoritmo que será parte de la llave de
            caché.
          maxLength: 50
        parameters:
          description: Diccionario dinámico de inputs o parámetros del job.
      required:
      - parameters
      - plugin_name
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
servers:
- url: http://localhost:8000
  description: Entorno local por hostname
- url: http://127.0.0.1:8000
  description: Entorno local por loopback
...